pyproject.toml
[project]
name = "hydra"
version = "0.1.0"
description = "Bitcoin auto-trading platform"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [{ name = "somatczk" }]

dependencies = [
    # Core
    "pydantic>=2.10,<3",
    "pydantic-settings>=2.7,<3",
    "pyyaml>=6.0,<7",
    "structlog>=24.4,<25",
    "msgpack>=1.1,<2",
    "redis[hiredis]>=5.2,<6",
    # Data
    "ccxt>=4.4,<5",
    "asyncpg>=0.30,<1",
    "alembic>=1.14,<2",
    "sqlalchemy[asyncio]>=2.0,<3",
    # Indicators
    "numpy>=2.1,<3",
    "pandas>=2.2,<3",
    "ta-lib>=0.5,<1; sys_platform != 'win32'",
    # ML
    "onnxruntime>=1.20,<2",
    # Backtest
    "vectorbt>=0.26,<1",
    # Execution
    "ccxt[async]>=4.4,<5",
    # Dashboard backend
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.34,<1",
    "websockets>=14.1,<15",
    # Telegram
    "python-telegram-bot>=21.9,<22",
    # Observability
    "prometheus-client>=0.21,<1",
]

[project.optional-dependencies]
ml-training = [
    "xgboost>=2.1,<3",
    "lightgbm>=4.5,<5",
    "scikit-learn>=1.6,<2",
    "torch>=2.5,<3",
    "optuna>=4.1,<5",
    "mlflow>=2.19,<3",
    "onnxmltools>=1.12,<2",
    "finrl>=0.3,<1",
]
dev = [
    "pytest>=8.3,<9",
    "pytest-asyncio>=0.25,<1",
    "pytest-cov>=6.0,<7",
    "pytest-mock>=3.14,<4",
    "pytest-benchmark>=5.1,<6",
    "pytest-timeout>=2.3,<3",
    "aioresponses>=0.7,<1",
    "factory-boy>=3.3,<4",
    "faker>=33.0,<34",
    "testcontainers[postgres,redis]>=4.9,<5",
    "ruff>=0.8,<1",
    "mypy>=1.13,<2",
    "bandit>=1.8,<2",
    "import-linter>=2.1,<3",
    "types-pyyaml>=6.0",
    "types-redis>=4.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hydra"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "unit: pure unit test, no infrastructure",
    "integration: requires TimescaleDB + Redis (testcontainers)",
    "e2e: full pipeline test, all infrastructure",
    "performance: benchmark test, may be slow",
    "slow: long-running test",
]
addopts = "-ra --strict-markers"

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "W",       # pycodestyle
    "F",             # pyflakes
    "I",             # isort
    "UP",            # pyupgrade
    "B",             # bugbear
    "SIM",           # simplify
    "N",             # pep8-naming
    "S",             # bandit
    "RUF",           # ruff-specific
]
ignore = ["S101", "S104"]  # allow assert in tests, binding to all interfaces

[tool.ruff.lint.isort]
known-first-party = ["hydra"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
exclude = ["tests/"]

[[tool.mypy.overrides]]
module = ["msgpack", "msgpack.*"]
ignore_missing_imports = true

[tool.coverage.run]
source = ["src/hydra"]
omit = ["*/migrations/*"]

[tool.coverage.report]
fail_under = 50
show_missing = true

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B104", "B324", "B608"]

[tool.importlinter]
root_packages = ["hydra"]

[[tool.importlinter.contracts]]
name = "Module boundaries"
type = "layers"
layers = [
    "hydra.dashboard",
    "hydra.portfolio | hydra.backtest",
    "hydra.execution | hydra.risk | hydra.ml",
    "hydra.strategy | hydra.data",
    "hydra.indicators",
    "hydra.core",
]
